+++ ChangeLog.md
@@ -7,6 +7,9 @@
   - Deprecate `Data.Barbie`, in favor of `Data.Functor.Barbie`.
   - Deprecate `Data.Barbie.Bare`, in favor of `Barbies.Bare`.
   - Deprecate `Data.Barbie.Constraints`, in favor of `Barbies.Constraints`.
+  - Create a `Barbies` module, to contain wrappers, basic docs, etc.
+    `Data.Functor.Barbie` contains only functor-related stuff.
+  - Add a `ErrorContainer` wrapper, similar to `Container` but for `Either e`.
 
 ## 1.1.3.0
   - `Wear` will raise a `TypeError` instead of getting